task_tracker_bot/requirements.txt
python-telegram-bot>=20.7
python-dotenv>=1.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
requests>=2.31.0
python-dateutil>=2.8.2

